Move logs fragment to fragment package
